requirements.txt
absl-py==1.4.0
accelerate==0.23.0
addict==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
antlr4-python3-runtime==4.9.3
anyio==3.7.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
async-lru==2.0.2
async-timeout==4.0.2
attrs==23.1.0
av==10.0.0
Babel==2.12.1
beautifulsoup4==4.12.2
bitsandbytes==0.39.1
bleach==6.0.0
blinker==1.6.2
blis==0.7.9
braceexpand==0.1.7
brotlipy==0.7.0
cachetools==5.3.1
catalogue==2.0.8
cfgv==3.3.1
click==8.1.4
confection==0.1.0
contexttimer==0.3.3
contourpy==1.1.0
# Editable Git install with no remote (cvar-pyutils==0.1.dev0)
cycler==0.11.0
cymem==2.0.7
datasets==2.13.1
decord==0.6.0
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
easydict==1.10
einops==0.6.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
evaluate==0.4.0
fairscale==0.4.4
fastjsonschema==2.17.1
filelock==3.12.2
fonttools==4.40.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.6.0
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
huggingface-hub==0.15.1
identify==2.5.24
imageio==2.31.1
importlib-metadata==6.7.0
inflect==7.0.0
iopath==0.1.10
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
joblib==1.3.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.2.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.23.0
kaggle==1.5.15
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.3
Markdown==3.4.4
markdown-it-py==3.0.0
matplotlib==3.7.1
mdurl==0.1.2
mistune==3.0.1
mkl-service==2.4.0
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nltk==3.8.1
nodeenv==1.8.0
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.26.0
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
overrides==7.3.1
pandas==2.0.2
pandocfilters==1.5.0
pathy==0.10.2
peft==0.5.0
# Editable install with no version control (peft==0.2.0.dev0)
Pillow==9.4.0
platformdirs==3.6.0
plotly==5.15.0
portalocker==2.7.0
pre-commit==3.3.3
preshed==3.0.8
prometheus-client==0.17.0
protobuf==4.23.4
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.6
pydantic==1.10.11
pydeck==0.8.1b0
pydevd-pycharm==213.7172.26
Pygments==2.15.1
Pympler==1.0.1
pyrsistent==0.19.3
python-json-logger==2.0.7
python-magic==0.4.27
python-slugify==8.0.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
qtconsole==5.4.3
QtPy==2.3.1
regex==2023.6.3
requests-oauthlib==1.3.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rouge==1.0.1
rouge-score==0.1.2
rsa==4.9
safetensors==0.3.1
-e git+https://github.com/salesforce/LAVIS.git@38f6a13eed6d98c64ae6a63a64f575c201c7aca6#egg=salesforce_lavis
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
-e git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588#egg=segment_anything
Send2Trash==1.8.2
sentencepiece==0.1.99
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
streamlit==1.24.1
supervision==0.6.0
tenacity==8.2.2
tensorboard==2.14.0
tensorboard-data-server==0.7.1
terminado==0.17.1
text-unidecode==1.3
thinc==8.1.10
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.4.12
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
-e git+https://github.com/huggingface/transformers.git@8edd0da960cb91f17e97803435a45887600d3984#egg=transformers
triton==2.0.0
typer==0.9.0
tzdata==2023.3
tzlocal==4.3.1
uri-template==1.2.0
validators==0.20.0
virtualenv==20.23.1
wasabi==1.1.2
watchdog==3.0.0
webcolors==1.13
webdataset==0.2.48
webencodings==0.5.1
websocket-client==1.6.0
Werkzeug==2.3.7
widgetsnbextension==4.0.7
xxhash==3.2.0
yapf==0.40.1
yarl==1.9.2
zipp==3.15.0
